map refactored
