+++ src/views/MapView.vue
@@ -15,30 +15,9 @@
                 <router-view @onShowInfoUser="showUser"></router-view>
             </div>
         </div>
-        <div class="cont--mapa">
-            <mapbox 
-                access-token="pk.eyJ1IjoiYW5nZWxyb2Rybzk1IiwiYSI6ImNqODljcTJrdDAxaWIyd21rNTZubHQwamMifQ.6ghwymwGfrRC15-iKOxcww"
-                :map-options="{
-                    style: 'mapbox://styles/mapbox/streets-v9',
-                    center: [-70.221799, -18.0031498],
-                    zoom: 15
-                }"
-                :geolocate-control="{
-                    show: true,
-                    position: 'top-left'
-                }"
-                :nav-control="{
-                    show: true, 
-                    position: 'top-left'
-                }"
-                @map-init="mapInitialized"
-                :fullscreen-control="{
-                    show: true,
-                    position: 'top-left'
-                }"
-            />
-           
-        </div>
+        <PostMap
+            @init="mapInitialized"
+        />
     </div>
 </template>
 
@@ -46,30 +25,21 @@
     /* eslint-disable */
     import { mapGetters, mapState } from "vuex";
     import {random} from "lodash";
-    import Mapbox from 'mapbox-gl-vue';
+    import mapMixin from "./common/map";
+
     export default {
         name: "Map",
+        mixins: [mapMixin],
         data() {
             return {
-                mapOptions: {
-                    zoom: 14,
-                    center: new google.maps.LatLng(-18.013611, -70.252769),
-                },
                 flagInfoUser: false,
-                crntUser: {},
-                map: {}
+                crntUser: {}
             }
-        },
-        components: {
-            Mapbox
         },
         watch: {
             markers() {
                 this.mapMarker();
             }
-        },
-        beforeDestroy() {
-            this.map.remove();
         },
         methods: {
             showUser(user) {
@@ -87,16 +57,6 @@
                   .addTo(this.map);
 
                 }
-            },
-            genLayoutMarker(data) {
-                const el = document.createElement("div");
-                el.className = (data.type == 1) ? "marker marker--encontrado" : "marker marker--perdido";
-                if (data.photo) {
-                    el.style.backgroundImage = `url(${data.photo})`;
-                }
-                el.style.width = data.properties.iconSize[0] ? (data.properties.iconSize[0] + 'px') : '48px';
-                el.style.height = data.properties.iconSize[1] ? (data.properties.iconSize[1] + 'px') : '48px';
-                return el;
             }
         },
         computed: {
